pyproject.toml
[tool.poetry]
name = "tscraper"
version = "0.1.0"
description = ""
authors = ["Maks K <cedabef@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.13"
telethon = "^1.38.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
pytest-mock = "^3.14.0"

[tool.poetry.scripts]
tscraper = "tscraper.tscraper:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
